+++ ui/litellm-dashboard/e2e_tests/fixtures/migratedPages.ts
@@ -10,16 +10,18 @@
  *
  * Keep this in lockstep with MIGRATED_PAGES in src/utils/migratedPages.ts.
  * Pending (add as each PR lands): the leaf-pages batch
- * (budgets, caching, cost-tracking, guardrails, guardrails-monitor, logs,
- * mcp-servers, memory, policies, prompts, search-tools, skills,
- * tag-management, tool-policies, transform-request, ui-theme, vector-stores,
- * workflows).
+ * (caching, cost-tracking, guardrails, logs, mcp-servers, memory,
+ * policies, prompts, search-tools, skills, tag-management, tool-policies,
+ * transform-request, ui-theme, vector-stores).
  */
 export const MIGRATED_E2E_PAGES: Record<string, string> = {
   api_ref: "api-reference",
   "llm-playground": "playground",
   projects: "projects",
   "access-groups": "access-groups",
+  budgets: "budgets",
+  workflows: "workflows",
+  "guardrails-monitor": "guardrails-monitor",
 };
 
 export const MIGRATED_E2E_SEGMENTS: string[] = [...new Set(Object.values(MIGRATED_E2E_PAGES))];
