+++ ui/litellm-dashboard/src/app/(dashboard)/guardrails-monitor/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import GuardrailsMonitorView from "./components/GuardrailsMonitorView";
+import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
+
+export default function GuardrailsMonitor() {
+  const { accessToken } = useAuthorized();
+  return <GuardrailsMonitorView accessToken={accessToken} />;
+}
