+++ ui/litellm-dashboard/src/app/(dashboard)/page.tsx
@@ -3,7 +3,6 @@
 import ModelsAndEndpointsView from "@/app/(dashboard)/models-and-endpoints/ModelsAndEndpointsView";
 import AdminPanel from "@/components/AdminPanel";
 import AgentsPanel from "@/components/agents";
-import BudgetPanel from "@/components/budgets/budget_panel";
 import CacheDashboard from "@/components/cache_dashboard";
 import ClaudeCodePluginsPanel from "@/components/claude_code_plugins";
 import { teamListCall as v2TeamListCall } from "@/app/(dashboard)/hooks/teams/useTeams";
@@ -12,7 +11,6 @@ import useProxySettings from "@/app/(dashboard)/hooks/proxySettings/useProxySett
 import LoadingScreen from "@/components/common_components/LoadingScreen";
 import { CostTrackingSettings } from "@/components/CostTrackingSettings";
 import GeneralSettings from "@/components/general_settings";
-import GuardrailsMonitorView from "@/components/GuardrailsMonitor/GuardrailsMonitorView";
 import GuardrailsPanel from "@/components/guardrails";
 import PoliciesPanel from "@/components/policies";
 import { Team } from "@/components/key_team_helpers/key_list";
@@ -37,7 +35,6 @@ import UserDashboard from "@/components/user_dashboard";
 import VectorStoreManagement from "@/components/vector_store_management";
 import ToolPoliciesView from "@/components/ToolPoliciesView";
 import { MemoryView } from "@/components/MemoryView";
-import WorkflowRuns from "@/components/workflow_runs";
 import SpendLogsTable from "@/components/view_logs";
 import ViewUserDashboard from "@/components/view_users";
 import { useAuth } from "@/contexts/AuthContext";
@@ -385,8 +382,6 @@ function CreateKeyPageContent() {
             <AdminPanel proxySettings={proxySettings} />
           ) : page == "logging-and-alerts" ? (
             <Settings userID={userID} userRole={userRole} accessToken={accessToken} premiumUser={premiumUser} />
-          ) : page == "budgets" ? (
-            <BudgetPanel accessToken={accessToken} />
           ) : page == "guardrails" ? (
             <GuardrailsPanel accessToken={accessToken} userRole={userRole} />
           ) : page == "policies" ? (
@@ -450,12 +445,8 @@ function CreateKeyPageContent() {
             <VectorStoreManagement accessToken={accessToken} userRole={userRole} userID={userID} />
           ) : page == "tool-policies" ? (
             <ToolPoliciesView accessToken={accessToken} userRole={userRole} />
-          ) : page == "workflows" ? (
-            <WorkflowRuns accessToken={accessToken} />
           ) : page == "memory" ? (
             <MemoryView accessToken={accessToken} userID={userID} userRole={userRole} />
-          ) : page == "guardrails-monitor" ? (
-            <GuardrailsMonitorView accessToken={accessToken} />
           ) : page == "new_usage" ? (
             <NewUsagePage teams={(teams as Team[]) ?? []} organizations={(organizations as Organization[]) ?? []} />
           ) : (
