+++ ui/litellm-dashboard/src/app/(dashboard)/workflows/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+
+import WorkflowRuns from "./WorkflowRuns";
+import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
+
+export default function Workflows() {
+  const { accessToken } = useAuthorized();
+  return <WorkflowRuns accessToken={accessToken} />;
+}
