+++ ui/litellm-dashboard/src/utils/migratedPages.ts
@@ -15,6 +15,9 @@ export const MIGRATED_PAGES: Record<string, string> = {
   "llm-playground": "playground",
   projects: "projects",
   "access-groups": "access-groups",
+  budgets: "budgets",
+  workflows: "workflows",
+  "guardrails-monitor": "guardrails-monitor",
 };
 
 function uiBase(): string {
